+++ app/src/main/java/com/example/snoker/view/main/fragment/MainFragment.kt
@@ -1,4 +1,4 @@
-package com.example.snoker.view.fragment
+package com.example.snoker.view.main.fragment
 
 import android.content.Context
 import android.os.Bundle
@@ -9,24 +9,24 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.snoker.R
 
-class ProgressFragment : Fragment() {
+class MainFragment : Fragment() {
     companion object {
-        const val TAG: String = "ProgressFragment"
+        const val TAG: String = "MainFragment"
 
-        fun newInstance() : ProgressFragment {
-            return ProgressFragment()
+        fun newInstance() : MainFragment {
+            return MainFragment()
         }
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        Log.d(TAG, "ProgressFragment On create Called")
+        Log.d(TAG, "MainFragment On create Called")
     }
 
     // 프레그먼트를 안고 있는 액티비티에 붙었을때
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        Log.d(TAG, "ProgressFragment on Attach Called")
+        Log.d(TAG, "MainFragment on Attach Called")
     }
 
     override fun onCreateView(
@@ -34,8 +34,8 @@ class ProgressFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        Log.d(TAG, "ProgressFragment on Create View Called")
+        Log.d(TAG, "MainFragment on Create View Called")
 
-        return inflater.inflate(R.layout.fragment_progress, container, false)
+        return inflater.inflate(R.layout.fragment_main, container, false)
     }
 }
