+++ app/src/main/res/layout/fragment_board.xml
@@ -6,74 +6,74 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-<!--    <ImageView-->
-<!--        android:id="@+id/topImage"-->
-<!--        android:layout_width="match_parent"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        android:background="@drawable/top"-->
-<!--        app:layout_constraintEnd_toEndOf="parent"-->
-<!--        app:layout_constraintStart_toStartOf="parent"-->
-<!--        app:layout_constraintTop_toTopOf="parent" />-->
+    <ImageView
+        android:id="@+id/topImage"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@drawable/top"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
-<!--    <ImageView-->
-<!--        android:id="@+id/logoImage"-->
-<!--        android:layout_width="wrap_content"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        android:layout_marginTop="32dp"-->
-<!--        android:layout_marginEnd="32dp"-->
-<!--        android:layout_marginRight="32dp"-->
-<!--        app:layout_constraintEnd_toEndOf="parent"-->
-<!--        app:layout_constraintTop_toTopOf="parent"-->
-<!--        app:srcCompat="@drawable/logo_10" />-->
+    <ImageView
+        android:id="@+id/logoImage"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="32dp"
+        android:layout_marginEnd="32dp"
+        android:layout_marginRight="32dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:srcCompat="@drawable/logo_10" />
 
-<!--    <ImageView-->
-<!--        android:id="@+id/settingImage"-->
-<!--        android:layout_width="11dp"-->
-<!--        android:layout_height="11dp"-->
-<!--        android:layout_marginEnd="4dp"-->
-<!--        android:layout_marginRight="4dp"-->
-<!--        app:layout_constraintBottom_toBottomOf="@+id/idText"-->
-<!--        app:layout_constraintEnd_toStartOf="@+id/idText"-->
-<!--        app:layout_constraintTop_toTopOf="@+id/idText"-->
-<!--        app:srcCompat="@drawable/icon_setting" />-->
+    <ImageView
+        android:id="@+id/settingImage"
+        android:layout_width="11dp"
+        android:layout_height="11dp"
+        android:layout_marginEnd="4dp"
+        android:layout_marginRight="4dp"
+        app:layout_constraintBottom_toBottomOf="@+id/idText"
+        app:layout_constraintEnd_toStartOf="@+id/idText"
+        app:layout_constraintTop_toTopOf="@+id/idText"
+        app:srcCompat="@drawable/icon_setting" />
 
-<!--    <TextView-->
-<!--        android:id="@+id/idText"-->
-<!--        android:layout_width="wrap_content"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        android:layout_marginStart="45dp"-->
-<!--        android:layout_marginLeft="45dp"-->
-<!--        android:layout_marginTop="32dp"-->
-<!--        android:fontFamily="@font/noto_sans_kr_bold"-->
-<!--        android:gravity="center_horizontal"-->
-<!--        android:includeFontPadding="false"-->
-<!--        android:letterSpacing="-0.06"-->
-<!--        android:lineSpacingExtra="0dp"-->
-<!--        android:text="분당요정"-->
-<!--        android:textColor="#ffffff"-->
-<!--        android:textSize="17sp"-->
-<!--        app:layout_constraintStart_toStartOf="parent"-->
-<!--        app:layout_constraintTop_toTopOf="parent" />-->
+    <TextView
+        android:id="@+id/idText"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="45dp"
+        android:layout_marginLeft="45dp"
+        android:layout_marginTop="32dp"
+        android:fontFamily="@font/noto_sans_kr_bold"
+        android:gravity="center_horizontal"
+        android:includeFontPadding="false"
+        android:letterSpacing="-0.06"
+        android:lineSpacingExtra="0dp"
+        android:text="분당요정"
+        android:textColor="#ffffff"
+        android:textSize="17sp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
-<!--    <TextView-->
-<!--        android:id="@+id/textView"-->
-<!--        android:layout_width="wrap_content"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        android:fontFamily="@font/noto_sans_kr_regular"-->
-<!--        android:textSize="17sp"-->
-<!--        android:textColor="#ffffff"-->
-<!--        android:letterSpacing="-0.06"-->
-<!--        android:includeFontPadding="false"-->
-<!--        android:text="무슨 생각을 하고 계시나요"-->
-<!--        app:layout_constraintStart_toStartOf="@+id/settingImage"-->
-<!--        app:layout_constraintTop_toBottomOf="@+id/idText" />-->
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:fontFamily="@font/noto_sans_kr_regular"
+        android:textSize="17sp"
+        android:textColor="#ffffff"
+        android:letterSpacing="-0.06"
+        android:includeFontPadding="false"
+        android:text="무슨 생각을 하고 계시나요"
+        app:layout_constraintStart_toStartOf="@+id/settingImage"
+        app:layout_constraintTop_toBottomOf="@+id/idText" />
 
-<!--    <androidx.recyclerview.widget.RecyclerView-->
-<!--        android:id="@+id/postList"-->
-<!--        android:layout_width="match_parent"-->
-<!--        android:layout_height="match_parent"-->
-<!--        android:layout_marginTop="8dp"-->
-<!--        app:layout_constraintTop_toBottomOf="@+id/topImage"-->
-<!--        tools:layout_editor_absoluteX="8dp" />-->
-
+    <ListView
+        android:id="@+id/postList"
+        android:layout_width="409dp"
+        android:layout_height="611dp"
+        android:layout_marginTop="8dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/topImage" />
 </androidx.constraintlayout.widget.ConstraintLayout>
